backend/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
channels==2.3.1
channels_redis